Started work on DemoInvite component, still lacks functionality
