+++ sg-frontend/src/components/Login/Login.js
@@ -3,14 +3,18 @@ import React from 'react'
 import AppContainer from '../presentational/AppContainer'
 import LoginTop from './LoginTop'
 import LoginBottom from './LoginBottom'
+import Footer from '../presentational/Footer'
 
 const Login = () => {
 
     return (
-        <AppContainer>
-            <LoginTop />
-            <LoginBottom />
-        </AppContainer>
+        <>
+            <AppContainer>
+                <LoginTop />
+                <LoginBottom />
+                <Footer />
+            </AppContainer>
+        </>
     )
 }
