+++ sg-frontend/src/components/MobileHome/MobileHome.js
@@ -1,6 +1,7 @@
 import React from 'react'
 
 import AppContainer from '../presentational/AppContainer'
+import Footer from '../presentational/Footer'
 import Menu from '../Menu/Menu'
 
 import MobileHomeBottom from './MobileHomeBottom'
@@ -13,6 +14,7 @@ const MobileHome = () => {
             <Menu />
             <MobileHomeTop />
             <MobileHomeBottom />
+            <Footer />
         </AppContainer>
     )
 }
