+++ sg-frontend/src/components/PasswordChange/PasswordChange.js
@@ -1,6 +1,8 @@
 import React from 'react'
 
 import AppContainer from '../presentational/AppContainer'
+import Footer from '../presentational/Footer'
+
 import PasswordChangeTop from './PasswordChangeTop'
 import PasswordChangeBottom from './PasswordChangeBottom'
 
@@ -10,6 +12,7 @@ const PasswordChange = () => {
         <AppContainer>
             <PasswordChangeTop />
             <PasswordChangeBottom />
+            <Footer />
         </AppContainer>
     )
 }
