+++ sg-frontend/src/components/ProfileDelete/ProfileDelete.js
@@ -1,6 +1,8 @@
 import React from 'react'
 
 import AppContainer from '../presentational/AppContainer'
+import Footer from '../presentational/Footer'
+
 import ProfileDeleteTop from './ProfileDeleteTop'
 import ProfileDeleteBottom from './ProfileDeleteBottom'
 
@@ -10,6 +12,7 @@ const ProfileDelete = () => {
         <AppContainer>
             <ProfileDeleteTop />
             <ProfileDeleteBottom />
+            <Footer />
         </AppContainer>
     )
 }
