+++ sg-frontend/src/components/SignUp/SignUp.js
@@ -1,6 +1,8 @@
 import React from 'react'
 
 import AppContainer from '../presentational/AppContainer'
+import Footer from '../presentational/Footer'
+
 import SignUpTop from './SignUpTop'
 import SignUpBottom from './SignUpBottom'
 
@@ -10,6 +12,7 @@ const SignUp = () => {
         <AppContainer>
             <SignUpTop />
             <SignUpBottom />
+            <Footer />
         </AppContainer>
     )
 }
